Change type of hashmap value to i32

+++ backend/src/db/models.rs
@@ -3,5 +3,5 @@ use std::collections::HashMap;
 #[derive(Deserialize, Serialize)]
 pub struct TimeData {
     pub category: String,
-    pub time_data: HashMap<String, u64>,
+    pub time_data: HashMap<String, i32>,
 }
